web/rest: moved rating plan prices from brand::/my/rating_plan_prices to client::/rating_plan_groups/{id}/prices
