+++ web/rest/brand/app/config/api/resources.yml
@@ -28,29 +28,6 @@ Model\Token:
             required: true
             description: 'Target brand admin user name'
 
-Model\RatingPlanPrices:
-  itemOperations: []
-  collectionOperations:
-    get_my_rating_plan_prices:
-      access_control: '"ROLE_BRAND_ADMIN" in roles'
-      method: 'GET'
-      path: '/my/rating_plan_prices'
-      route_name: 'get_my_rating_plan_prices'
-      parameters: []
-      swagger_context:
-        produces:
-          - 'text/csv'
-        tags:
-          - My
-        parameters:
-          - name: id
-            in: query
-            type: integer
-            required: true
-            description: 'Rating plan group Id'
-  attributes:
-    pagination_client_enabled: true
-
 Model\ActiveCalls:
   itemOperations:
     get_my_active_calls:
