+++ web/rest/brand/app/config/routing.yml
@@ -30,15 +30,6 @@ gesdinet_jwt_refresh_token:
 ###############################
 ## Custom routes
 ###############################
-get_my_rating_plan_prices:
-  path: '/my/rating_plan_prices'
-  methods: ['GET']
-  defaults:
-    _controller: Controller\My\RatingPlanPricesAction
-    _api_resource_class: 'Model\RatingPlanPrices'
-    _api_collection_operation_name: 'get_my_rating_plan_prices'
-    _api_receive: false
-
 get_my_active_calls:
   path: '/my/active_calls'
   methods: ['GET']
