+++ web/rest/client/app/config/routing.yml
@@ -30,6 +30,15 @@ gesdinet_jwt_refresh_token:
 ###############################
 ## Custom routes
 ###############################
+get_my_rating_plan_prices:
+  path: '/rating_plan_groups/{id}/prices'
+  methods: ['GET']
+  defaults:
+    _controller: Controller\Provider\RatingPlanPricesAction
+    _api_resource_class: 'Model\RatingPlanPrices'
+    _api_collection_operation_name: 'get_my_rating_plan_prices'
+    _api_receive: false
+
 get_my_status:
   path: '/my/status'
   methods: ['GET']
