+++ web/rest/client/features/provider/ratingPlanGroup/getPrices.feature
